+++ mangaScraper/Support/fetcher.js
@@ -1,5 +1,13 @@
 
 export class Fetcher {
+    constructor() {
+        this.config = {};
+    }
+    updateConfig(config) {
+        this.config = config;
+        return;
+    }
+
     async site(link) {
         /**
          * Fetches the content of a website and returns the parsed HTML document, or null if failed.