+++ mangaScraper/mangaScraper.js
@@ -8,7 +8,26 @@ import { Mangatoto }  from './Fetchers/mangatoto.js';
 
 
 export class MangaSearch {
-    static search() {
+    constructor() {
+
+        this.requesthandler = new RequestHandler();
+
+        this.config = {
+
+            askRound : 0
+        }
+    }
+
+    updateConfig(config = {}) {
+        config = this.merge.info(this.config, config);
+
+        this.config = config;
+        
+        this.requesthandler.updateConfig(config);
+        return;
+    }
+
+    static search(query, askRound = this.config.askRound, callback) {
         // Implement MangaSearch search method if needed
     }
 }
@@ -24,20 +43,25 @@ export class Manga {
         
         // Config stores all default function params
         this.config = {
-            // Genric
-            language : null,
-            chapter : 0,
+            // Manga
+             // all
+             language : null,
+             chapter : 0,
+             // get()
+             outputAll : false,
+             outputSource : false,
+            
+            // RequestHandler
+             // orderRequests()
+             maxParallelRequests : 2,
 
-            // Update
-            maxParallelRequests : 2,
-
-            // Get
-            outputAll : false,
-            outputSource : false,
-            fallbackLanguage : true,
-            alwaysOutput : true,
-            justChapter : false
+            // InfoSourceHelper
+             // getItems
+             fallbackLanguage : true,
+             alwaysOutput : true,
+             justChapter : false
         }
+        this.updateConfig();
         // INFO
         
         this.infoSources = [];
@@ -80,9 +104,20 @@ export class Manga {
          *     ]
          */
     }
+    
+    updateConfig(config = {}) {
+        config = this.merge.info(this.config, config);
 
+        this.config = config;
 
-    async update(items, language = this.config.language, chapter = this.config.chapter, maxParallelRequests = this.config.maxParallelRequests) {
+        this.infoSourceHelper.updateConfig(config);
+        this.requesthandler.updateConfig(config);
+        this.templater.updateConfig(config);
+        this.merge.updateConfig(config);
+        return;
+    }
+
+    async update(items, language = this.config.language, chapter = this.config.chapter) {
         /**
          * returns nothing. Just updates the value of the item requested
          * @param {Array} item - list of names of the items you want to get update
@@ -141,28 +176,11 @@ export class Manga {
                  * This is intentional although I may add a flag/setting to merge.
                  */
 
-                let rawRequestOrder = JSON.parse(JSON.stringify(this.sourceRank)); // This is done to make a deep copy of sourceRank
-                let requestOrder = [];
-
-                for (let rank of rawRequestOrder) {
-
-                    let requestItem = [];
-                    while(rank.length !== 0){
-
-                        if (requestItem.length >= maxParallelRequests) {
-                            requestOrder.push(requestItem);
-                            requestItem = [];
-                        }
-                        requestItem.push(rank[0]);
-                        rank.shift()
-                    }
-
-                    requestOrder.push(requestItem);
-                }
+                let requestOrder = this.requesthandler.orderRequests(this.sourceRank);
 
                 //Make request
                 for (let requestGroup of requestOrder) {
-                    let parallelRequestsOut = await this.requesthandler.parallelizeRequests(item, language, chapter, requestGroup, this.infoSources);
+                    let parallelRequestsOut = await this.requesthandler.parallelizeUpdateRequests(item, language, chapter, requestGroup, this.infoSources);
                     if (parallelRequestsOut !== null) {
                         this.infoSources = parallelRequestsOut;
                         if (this.infoSourceHelper.countItem(item, language, chapter, this.infoSources) !== 0) {
@@ -286,14 +304,46 @@ class RequestHandler{
         // Fetchers
         this.mangatoto = new Mangatoto();
 
+        this.config = {};
     }
-    async distributeRequest(item, language, chapter, source, info) {
-        // Stage 0 create useful vars
+
+    updateConfig(config) {
+        config = JSON.parse(JSON.stringify(config));
+
+        this.config = config;
+
+        this.mangatoto.updateConfig(config);
+        return;
+    }
+
+    orderRequests(sourceRank) {
+        let rawRequestOrder = JSON.parse(JSON.stringify(sourceRank)); // This is done to make a deep copy of sourceRank
+        let requestOrder = [];
+
+        for (let rank of rawRequestOrder) {
+
+            let requestItem = [];
+            while(rank.length !== 0){
+
+                if (requestItem.length >= this.config.maxParallelRequests) {
+                    requestOrder.push(requestItem);
+                    requestItem = [];
+                }
+                requestItem.push(rank[0]);
+                rank.shift()
+            }
+
+            requestOrder.push(requestItem);
+        }
+        return requestOrder;
+    }
+
+
+    async distributeUpdateRequest(item, language, chapter, source, info) {
+        // Create useful vars
         let dashIndex = source.indexOf('-');
         let rawID = dashIndex !== -1 ? source.substring(dashIndex + 1) : source;
         let rawSource = dashIndex !== -1 ? source.substring(0, dashIndex) : source;
-
-        // Stage 1 call module functions
 
         // Mangatoto.js
         if (rawSource == this.mangatoto.source) {
@@ -313,7 +363,28 @@ class RequestHandler{
         return info;
     }
 
-    async parallelizeRequests(item, language, chapter, requestGroup, infoSources) {
+    async distributeSearchRequest(query, askRound, source) {
+        // Stage 0 create useful vars
+        let dashIndex = source.indexOf('-');
+        let rawID = dashIndex !== -1 ? source.substring(dashIndex + 1) : source;
+        let rawSource = dashIndex !== -1 ? source.substring(0, dashIndex) : source;
+
+        let result;
+
+        // Mangatoto.js
+        if (rawSource == this.mangatoto.source) {
+            result = this.mangatoto.search(query, askRound);
+        }
+
+        // We throw error here to stop it from being added to 
+        if (result == null) {
+            throw new console.error("Variable 'result' cannot be null.");
+        }
+
+        return result;
+    }
+
+    async parallelizeUpdateRequests(item, language, chapter, requestGroup, infoSources) {
         if (!Array.isArray(requestGroup)) {
             console.error("Invallid Input: requestGroup is not list");
             return null;
@@ -328,7 +399,7 @@ class RequestHandler{
             const requestPromise = new Promise((resolve, reject) => {
                 // Make the request
                 let info = this.infoSourceHelper.getInfo(requestGroup[i], infoSources);
-                this.distributeRequest(item, language, chapter, requestGroup[i], info)
+                this.distributeUpdateRequest(item, language, chapter, requestGroup[i], info)
                     .then(response => {
                         resolvedRequests.push(requestGroup[i]);
                         resolve(response); // Resolve the promise with the response