+++ mangaScraper/Docs/ChangeLog.md
@@ -1,6 +1,7 @@
 # V 0.4
  - Added the updateConfig function ([Documentation here](API.md##UpdateConfig)) to replace customizable defaults added in V0.1.4
  - Added a function to all classes called updateConfig()
+ - Added orderRequests() to RequestHandler
  - Added Documention for the new changes
  - Renamed functions in RequestHandler
 ## V 0.3.4